refactor: main.js 모듈 분리 (3549줄 → 9개 파일)
단일 main.js를 논리적 단위 9개 파일로 분리:
- constants.js: 캔버스·상수·적 타입 정의
- towers.js: TOWER_TYPES, TOWER_ORDER
- utils.js: 유틸리티 함수 (그리드·색상·수학·포탑 계산)
- map.js: 맵 정의·경로 데이터·buildMapData
- ui.js: 게임 상태·DOM 참조·UI 관리 함수
- audio.js: Web Audio 엔진
- game.js: 게임 로직·전투·update 루프
- renderer.js: 렌더링 파이프라인·render 함수
- main.js: 이벤트 바인딩·게임 루프·초기화

전역 변수 공유 방식 유지 (ES Module 미사용).
index.html script 태그 9개로 교체.
테스트 vm.runInThisContext 방식으로 전환.

Closes #1
