+++ index.html
@@ -130,6 +130,14 @@
 
     <div id="a11y-announcer" class="sr-only" aria-live="assertive" aria-atomic="true"></div>
 
+    <script src="constants.js"></script>
+    <script src="towers.js"></script>
+    <script src="utils.js"></script>
+    <script src="map.js"></script>
+    <script src="ui.js"></script>
+    <script src="audio.js"></script>
+    <script src="game.js"></script>
+    <script src="renderer.js"></script>
     <script src="main.js"></script>
 </body>
 </html>
