+++ tests/smoke.test.js
@@ -1,6 +1,7 @@
-﻿const { JSDOM } = require('jsdom');
+const { JSDOM } = require('jsdom');
 const fs = require('fs');
 const path = require('path');
+const vm = require('vm');
 
 function noop() {}
 
@@ -100,8 +101,22 @@ function setupDom() {
         drawImage: noop
     });
 
-    delete require.cache[require.resolve('../main.js')];
-    require('../main.js');
+    const scriptFiles = [
+        'constants.js',
+        'towers.js',
+        'utils.js',
+        'map.js',
+        'ui.js',
+        'audio.js',
+        'game.js',
+        'renderer.js',
+        'main.js'
+    ];
+    for (const file of scriptFiles) {
+        const filePath = path.join(__dirname, '..', file);
+        const code = fs.readFileSync(filePath, 'utf-8');
+        vm.runInThisContext(code, { filename: filePath });
+    }
 
     return { window, document };
 }
